services/api/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
supabase==2.6.0
httpx==0.27.2
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0
polars==0.20.3
numpy==1.26.3
google-generativeai==0.7.2
PyJWT==2.8.0
websockets==12.0
pytest==7.4.4
cryptography==42.0.5
